User Friendly Exception Handling Behaviour
The purpose of this change are:

* Prevent stacktrace to be displayed when operation is cancelled with `Ctrl+C`
* When internal error occurs, display a user-friendly error message instead of stacktrace, and save the stacktrace into a separate log file.

Required changes are:

* All the main functionality are moved into its own modules. This means `conduct.py`, `sandbox.py`, and `shazar.py` now has its own equivalent `conduct_main.py`, `sandbox_main.py`, and `shazar_main.py` respectively.
* The main files - `conduct.py`, `sandbox.py`, and `shazar.py` - are responsible for loading and running its respective main module.
* The main files are responsible for handling exceptions that may be raised by the main module, i.e. display friendly error message and log stacktrace into the file
* With this setup, the main files are able to wrap the loading and execution of the main module within a `try/except` block.
* This means exception encountered when loading the module will still have a friendly treatment. This is difficult to achieve without moving the main functionality into its own module.
* The main files has a special treatment for `Ctrl+C` (i.e.`KeyboardInterrupt`) and system exit (i.e. `SystemExit`)
* The functionality of this exception handling are consolidated in `main_handler.py` module which is used by the main files.
